Emblem For (Encrypted) Volume Icons
GLib may give emblems to volume icons but drawing emblems was not possible because there is no size hint in PlacesModelItem (see #41 (comment)). This PR simply gets the required icon size from PlacesView (through PlacesModel) and draws the GIcon emblem.

In my opinion, the hard task of hacking XdgIcon is an overkill for such a simple job -- we don't need emblems in other places (for now).