Move fixes (#70)
* terminal: Allow copy elision (NRVO)

<cite>
If a function returns a class type by value, and the return statement's
expression is the name of a non-volatile object with automatic storage
duration, which isn't the function parameter, or a catch clause
parameter, and which has the same type (ignoring top-level
cv-qualification) as the return type of the function, then
copy/move (since C++11) is omitted. When that local object is
constructed, it is constructed directly in the storage where the
function's return value would otherwise be moved or copied to. This
variant of copy elision is known as NRVO, "named return value
optimization".
</cite>

* Avoid use of invalid objects after "move"

* dirtreemodelitem: Use end() properly

..and remove the double check for the end-ness.

* volumemanager: Do not create (unneeded) temp object

* FileSearchDialog: Do not use std::move()

..as the object returned from function is a perfectly valid rvalue.