Respect inactiveness when drawing text
Fixes https://github.com/lxde/pcmanfm-qt/issues/560.

Previously, the text was always drawn with the active (normal) color group, even when the state was not active. This patch sets the correct color group for both normal and selected texts. Although the difference is only visible with Qt styles (like Kvantum) that handle inactiveness, the logic should be respected.