Smooth scrolling for icon and thumbnail views
Fixes https://github.com/lxde/pcmanfm-qt/issues/579.

The smooth wheel scrolling is implemented when the cursor is inside the icon/thumbnail view. The scrolling is as before (jumpy) when the cursor is on the scrollbar, so that both options are available to the user.

The CPU usage is optimized by disabling the animation when the wheel speed is too high. Unlike in Dolphin, the animation respects the number of scroll lines per wheel turn, as set in the LXQt configuration.