Fix wrong gray-out of cut files
Fixes https://github.com/lxde/pcmanfm-qt/issues/596.

The selected files of a folder are cut only if their paths are equal to the cut paths. Otherwise, the selection shouldn't be used to gray out items. Previously, the selection was always used to gray out items and that resulted in a wrong gray-out when a folder was opened in more that one tab/window.