Prevent a potential crash in xdndworkaround
The pointer `lastDrag_`, which is used by `xdndworkaround.cpp` to kepp track of the last drag, may become dangling in the meanwhile (when used in `dndWorkaround::selectionRequest`) and cause a crash. To prevent that, `QPointer` is used.

NOTES:
(1) Without tab DND, there was no chance for `lastDrag_` to become a dangling pointer. However, `QPointer` is always safer.
(2) Before this patch, I could reproduce a crash only under Enlightenment (it's a good environment for finding hidden issues because of its unusual ways of handling X).