Prepare libfm-qt for bulk rename
This will be used and followed by a PR for bulk rename in pcmanfm-qt. `changeFileName()` returns `true` only if renaming is successful and shows an error message only if needed. The changes are backward compatible.

NOTE: Bulk rename is not included in libfm-qt itself because only a file manager may need it.