Fullscreen shortcut for file dialog
Fixes https://github.com/lxqt/libfm-qt/issues/244

NOTE: saveGeometry/restoreGeometry isn't added to the code because we don't want to restore the position and, especially, we don't want the dialog to be opened on another screen. So, as before, only the size is saved but `F11` is added as the fullscreen shortcut.