Fixed the logic of selection corners
Although no bug is found in selection corners, their logic had a serious flaw, which is fixed by this simple patch.

In `FolderViewListView::indexAt()`, the mutable variable `cursorOnSelectionCorner_` should be set only when the point is the cursor position. The reason is that, on the one hand, that variable is used only for the hover state and, on the other hand, `indexAt()` may be called regardless of the cursor position.

Context: I'd succeeded in sabotaging selection corners by calling `indexAt()` from inside Kvantum ;) Only after that, I saw the problem. Strange that selection corners worked before!