Fixed mimetype filters in LXQt file dialog
Fixes https://github.com/lxqt/libfm-qt/issues/342

(1) The problem was a just typo, which is corrected.

(2) I also took the liberty of replacing `QRegExp` with `QRegularExpression` because the former is deprecated -- I know it's a bad practice to have two changes in one PR but I might never do it if not here.

NOTE: The only way of testing this is using `QFileDialog::setMimeTypeFilters()` in a code. It passed my tests though.