Prevented a rare random crash in file dialog
On selecting a file path, file dialog also scrolls to it with a single-shot timer and a lambda. Previously, the file index was captured by the lambda but, under very rare circumstances, if the directory had been changed when the slot was called, the index would be irrelevant and so, a crash would happen. This patch captures the file path instead.

Although the crash was rare, sometimes I was able to reproduce it by:

(1) Opening file dialog from FeatherPad or Kate;
(2) Selecting a folder and then a file inside it with `Ctrl` + mouse; and
(3) Pressing `Enter`.