Always activate a selected item with Enter/Return
Closes https://github.com/lxqt/pcmanfm-qt/issues/913

When `Enter` is pressed, if the current item isn't selected or if there's no current item, activate the first selected item instead.

Not activating a selected item was a counterintuitive behavior (and, for example, never happens in Dolphin); the user saw a selection without being able to do anything with it by pressing Enter/Return.

NOTE: With recent patches for pcmanfm-qt and libfm-qt, this means that all selected items will be lauched in pcmanfm-qt on pressing Enter/Return, whether one of them is the current item or not.