Realod folder after transfer job if it lacks file monitoring
Closes https://github.com/lxqt/pcmanfm-qt/issues/933 and closes https://github.com/lxqt/libfm-qt/issues/280

After a file transfer job is finished inside a directory, if it is the path of an open folder that lacks file monitoring, this patch reloads its corresponding folder. In this way, the lack of file monitoring is partially compensated for.

Please note that this doesn't work with `search://` because the files inside `search://` don't belong to it.

WARNING: All libfm-qt based apps should be recompiled after this.