Disable selection coners in single-selection mode
Fixes https://github.com/lxqt/libfm-qt/issues/414

For example, in lximage-qt's thumbnail view, selection corners not only caused a wrong behavior but also activated the rubber-band selection. The same was true for LXQt file dialog in the single-selection mode.

This patch enables the selection corners only for extended and multiple selection modes and then treats the extended mode as the default one.

It also disables rubber-band selection in tree-view when the latter is in the single-selection mode.