Row-by-row scrolling inside view when Shift is pressed
Some users of Qt apps know that Shift may change wheel scrolling speed. This patch makes the view scroll row-by-row with the mouse wheel when Shift is pressed and the cursor is inside the view.

Of course, in the compact view, Shift wheel scrolling is done column-by-column.

When the cursor is on the scrollbar, Qt's default behavior is left intact, i.e. the scroll is faster with Shift. Although I don't think users of a file manager need that, I didn't add extra code to change it.

NOTE: The arguments of `QWheelEvent` have changed recently. I used the old constructor to not break the compilation with older Qt versions but we should really bump the minimum version.