Added deletion time to Trash
Outside Trash, the deletion column and sort action are hidden/removed because the column is empty and, if sorting is done by deletion inside Trash, it will be by file name outside it. That's unlike Dolphin, which shows a redundant empty column outside Trash.

NOTE: pcmanfm-qt should be recompiled, preferably with the patch that will follow this for a complete implementation.