Don't start item DND until the cursor is really moved
Closes https://github.com/lxqt/pcmanfm-qt/issues/976

Qt considers the cursor position relative to the view to determine whether a DND should start, so that, even when the cursor isn't moved, the DND will start if the view is scrolled (by mouse wheel).

However, the user may want to press the left mouse button on an item and scroll the view with mouse wheel before starting DND. This patch makes that possible by preventing a DND until the global cursor position is changed.

NOTE: Cursor position relative to the window might also work but the global position seems good (for now).