Fixed problems in graying-out of cut files
Fixes https://github.com/lxqt/libfm-qt/issues/457

Cut files are grayed out almost instantly once the signal `QClipboard::dataChanged` is emitted, even when their number is huge or they're cut by another app (based on libfm-qt or not).

The structure of the graying-out code is simplified, so that its process is started in `Fm::FolderModel` by connecting to `QClipboard::dataChanged` but the necessary updates are done by `Fm::Folder::updateCutFiles()`, which is very fast.

Please note that, when cutting is done in *another* app, `QClipboard::dataChanged` isn't always reliable: the signal may be emitted with a delay or may not be emitted until the cursor enters the window. That's a Qt issue and is especially visible in Qt 5.13 but it isn't so important for us because we want it to work with different folder models of the same app.

NOTE: All libfm-qt based apps should be recompiled.