An option to show folder context menu with Ctrl + right click
If it's enabled (by default, it's disabled), when the user right clicks anywhere inside the view while pressing Ctrl, all items will be deselected and the folder context menu will be shown. In this way, there is no need to search for an empty space.

A pcmanfm-qt PR will follow this to make an option out of it.

NOTE: Recompile libfm-qt based apps! pcmanfm-qt is better to be recompiled with the next patch.