Enforce eject button width (on side pane)
Because of a bug since Qt 5.11, `QTreeView::setColumnWidth` doesn't work as expected when the minimum section width isn't set in the code. Apparently, Qt devs see it as a "feature" (https://bugreports.qt.io/browse/QTBUG-68503) but, IMO, it's a bug because it breaks a method — as it did in our code and wasted my time :(

Also, set the margins of the eject button by consulting the style — and not by using a fixed value (a mistake many coders make).

Closes https://github.com/lxqt/pcmanfm-qt/issues/1018