Corrected path button "hidden" names
Each path button has 2 names, one of which is shown on it and the other one is used to get the full path out of its components (between slashes). In most cases, these strings are identical but in the case of trash:/// and maybe some other places, they may be different. This patch, gets the hidden name in such a way that it can be used to remake the full path correctly.

Fixes https://github.com/lxqt/pcmanfm-qt/issues/1020