Never consider icons selected in icon/thumbnail view
This patch removes the `Selected` state from icons in icon/thumbnail view because they never have the selection background: in icon/thumbnail view, we select texts, not icons.

Previously, the code didn't consider that in handling states. As a result, an SVG symbolic icon might change color and have a low contrast with its background when its item was selected.