Better focus change with Tab/BackTab in file dialog
With this patch, Tab key switches the focus from the main view to the name entry and BackTab does the reverse.

A filter event is used because `QWidget::setTabOrder()` cannot set the desired tab order in the file dialog (probably, due to an old Qt bug).

Closes https://github.com/lxqt/libfm-qt/issues/557

NOTE: Recompile pcmanfm-qt and lxqt-archiver after this.