Better code for graying out cut files
The current patch simplifies and enhances the graying-out of cut files considerably by removing all traces of cut files from the core of libfm-qt and doing the job at the level of `FolderModel`. In this way, a better functionality is achieved by a much simpler logic.

Since a hash set is used for tracking cut files (instead of a path vector), the CPU usage is negligible even with huge numbers of cut files.

In addition:

 * The patch grays out cut files in multiple folders too (e.g., when multiple files are cut from a recursive search result).
 * Cut files are also grayed out in folders without file monitor (except for `search://`, which isn't a real folder).

WARNING: The changes affect both ABI and API. Therefore, **pcmanfm-qt and all other libfm-qt based apps should be cleanly recompiled after this.**

Closes https://github.com/lxqt/libfm-qt/issues/563