Also Update cut files when entering a loaded folder
Because the folder may have been loaded with a separate `FolderModel` in the same app.

The following bug is fixed:

 1. Open a single pcmanfm-qt window and change directory to the desktop.
 2. Cut a file in the desktop from inside the window.
 3. Close the window, open it and change directory to the desktop again.

Then the cut file won't be grayed out without this fix (until the folder is reloaded) because the desktop is loaded in the background with a separate `FolderModel` (without `CachedFolderModel`).