Set a separate file size limit for external thumbnails
Since we don't create external thumbnails and their corresponding apps are responsible for them, we shouldn't force the file size limit of the built-in thumbnailer on them.

By default, the external limit is -1 (= no limit), while the internal limit is 4 MiB.

WARNING: All libfm-qt based apps should be recompiled after this.

NOTE: A PR for pcmanfm-qt will follow this to add an option for external thumbnailers.

Closes https://github.com/lxqt/libfm-qt/issues/642