Fixed blurred thumbnails with scale factors > 1
The device pixel ratio of the thumbnail image is set to that of the app and its size is multiplied by it, so that the thumbnail has the correct size and is as sharp as possible with scaling.

Closes https://github.com/lxqt/libfm-qt/issues/655

NOTE: The device pixel ratio of `QWindow` cannot be used in the code because "we don't know which window we are targeting" (quoted from Qt doc → QGuiApplication Class).