Workaround for a bug is GFileMonitor regarding symlinks (#672)
If a single `GFileMonitor` is unreffed, those of open folders with the same (symlink) target will stop emitting signals and become useless. That's an old bug in GLib and this patch is a workaround for it by:

 1. Unreffing the old file monitor only after a new one is created when reloading a folder; and
 2. Finding and reloading all open folders that have the same target of the folder that is closed.

A new method is added to `Fm::FileInfo` to get the file identifier. By using a simple pointer comparison, it can be used to find files that have the same target.

Closes https://github.com/lxqt/pcmanfm-qt/issues/607

NOTE: All libfm-qt based apps should be recompiled.