Fixed device usage of file properties dialog is rare cases
The device usage of a folder (like `/tmp`) may be different from that of its parent folder. In such cases, the file properties dialog showed different device usages on right clicking an empty space of the folder, depending on whether the parent folder was also opened (in another tab) or not.

This patch is rather a compromise between a total fix, which would require changes to the codes of `libfm-qt`-based apps, and the current situation. It practically fixes the issue without needing code changes to those apps.

In addition, a small mistake in the old code is corrected: instead of finding opened folders, the old code tried to load folders.

WARNING: All `libfm-qt`-based apps should be recompiled; otherwise they might not run after applying this patch (and will give "symbol lookup error").

Fixes https://github.com/lxqt/libfm-qt/issues/696