Fixed device usage of file properties dialog (is rare cases)
The device usage of a folder (like `/tmp`) may be different from that of its parent folder. In such cases, the file properties dialog showed different device usages on right clicking an empty space inside the folder, depending on whether the parent folder was also opened (in another tab) or not.

This patch fixes the issue by examining the folder and its parent.

Fixes https://github.com/lxqt/libfm-qt/issues/696