Avoided a freeze on mounting encrypted volumes
For some reason unknown to me (GVFS or Qt bug?), having any local event loop in response to `GMountOperation`'s signal "ask-password" will cause a total app freeze if the mounting is done by `g_volume_mount()` *and* a volume is added/removed during the loop. Such a loop is needed for entering the password. The freeze also happens on auto-mounting a drive with both encrypted and non-encrypted partitions, for the same reason.

This patch works around the problem by *not* showing the password dialog in response to "ask-password" when `g_volume_mount()` is used. Instead, it first tries to mount the volume without a password. In case a password is needed, (a) it shows the password dialog only after the mount failure but without showing an error message, (b) saves the results of the user's actions and (c) mounts the volume again by using those results.

To the user, everything should seem like before but without the possibility of a freeze; he doesn't see the stages a, b and c.

NOTE: All `libfm-qt`-based apps should be recompiled.

Closes https://github.com/lxqt/pcmanfm-qt/issues/1414