Added option for smooth scrolling to LXQt file dialog
It appears only with list and compact modes. By default, smooth scrolling is enabled.

The reason behind this option is that Qt may slow down item sorting when the number of items is huge and scrolling is done per pixel in list or compact mode. As a workaround, the user can disable per-pixel scrolling by disabling smooth scrolling.

Closes https://github.com/lxqt/libfm-qt/issues/739

WARNING: All libfm-qt based apps should be recompiled.