Allow remembering the result of execution prompt for multiple files
When multiple executable files are going to be launched, the execution prompt dialog gets a checkbox for remembering the user's action for next files of the same kind.

If you've selected 10 desktop entries inside `/usr/share/applications` and pressed `Enter` by mistake, you know why this is necessary.

WARNING: `pcmanfm-qt` and other apps that depend on `libfm-qt` should be recompiled after applying this patch.

Closes https://github.com/lxqt/pcmanfm-qt/issues/1616