Fixed unresponsive DND menu under Wayland
Under Wayland, not only the DND menu was unresponsive to mouseover, it might also cause a crash if Escape was pressed (twice).

The reason was that the menu was shown synchronously, before the DND was finished. This patch fixes the issue by showing the menu after finishing the DND (which is OK under X11 too).

Also, corrected the position of the side-pane DND menu under Wayland.

NOTE: Avoiding `QMenu::exec()` and using `QMenu::popup()` with signals and slots would be a cleaner solution. However, the current solution is much simpler and works fine. At some point, I might change the code to avoid `QMenu::exec()` as far as possible.

Fixes https://github.com/lxqt/libfm-qt/issues/863