Get terminals list also from data directories (#924)
The files `terminals.list` inside data directories take priority over the one provided by `libfm-qt` if existing. How data directories are handled is left to Glib.

In this way, the user can create `~/.local/share/libfm-qt/terminals.list` and see the result in PCManFM-Qt on the fly. (Later, PCManFM-Qt might be patched to create or append to that file.)

Closes https://github.com/lxqt/libfm-qt/issues/922