Started Qt6 port (#778)
* Started Qt6 port

* Focus the text entry of new-file dialog

Qt did it before Qt 6.6.1. Now it needs to be done explicitly.

* Also focus the text entry of overwrite dialog explicitly

Qt did it before Qt 6.6.1. Now it needs to be done explicitly.

I think this is a regression in Qt 6.6.1: the first focusable widget should be focused automatically, as it was always so. But the workaround is a harmless short line; hence adding it.

* Removed `qAsConst`

And bumped minimum required version of Qt to 6.6.0

* Silence new warnings of GLib ≥ 2.76

Apparently, GLib has become more strict; hence the warnings.