Allow paths in name entry of file dialog (#964)
A path can be absolute or relative and can be used with file opening and saving alike.

Also, two minor issues are revealed and fixed after this change:

 * A mistake is corrected in the logic of showing error message about nonexistent files.
 * An issue is fixed in updating accept button on changing directory.

NOTE: Although this isn't self-consistent, Qt's default file dialog supports it. It can also be useful when the path is already copied from somewhere. The extra code is kept to a minimum.

Closes https://github.com/lxqt/libfm-qt/issues/962