Enabled the setting about local-only thumbnails
The code relies on the attribute `G_FILE_ATTRIBUTE_FILESYSTEM_REMOTE`; whether it's set correctly or not is the responsibility of GLib/GVFS.

Closes https://github.com/lxqt/libfm-qt/issues/643

NOTE1: To be on the safe side, recompile all apps that are based on libfm-qt — but I didn't :P

NOTE2: I haven't seen an external thumbnailer that works with remote files. So, this is mostly about libfm-qt's internal thumbnailer.