Use our own bookmarks file
The new file is `~/.local/share/libfm-qt/bookmarks`. If it doesn't exist or is empty, as soon as a libfm-qt based app is launched, the GTK3 file `~/.config/gtk-3.0/bookmarks` will be consulted for backward compatibility, then the new file will be created and the GTK3 file will be ignored.

Closes https://github.com/lxqt/libfm-qt/issues/1009

NOTE: The portal file dialogs will see the change after a reboot (or after restarting user portal services).