Detect remote and read-only file systems (#1007)
The patch does it only by a single info query for the parent directory. So, it shouldn't affect the speed.

Also, the setting about local-only thumbnails is enabled in this way.

Closes https://github.com/lxqt/libfm-qt/issues/643 and fixes https://github.com/lxqt/libfm-qt/issues/1006

NOTE: To be on the safe side, recompile libfm-qt based apps — although the LXQt apps don't seem to need that.

REMINDER: Add an option to lximage-qt for local-only thumbnails if this PR is merged.