Enabled custom options for local terminals list
The `custom_args` key is read only from the local list (the global list doesn't — and shouldn't — have it), while `open_arg` and `noclose_arg` are first read from the local list, and if empty, from the global list.

Also, a memory leak and a bug are fixed in `terminal.cpp`.

Closes https://github.com/lxqt/libfm-qt/issues/954

NOTE: This PR will be followed by a pcmanfm-qt PR that uses it.